+++ src/core_modules/capture-core/components/ListView/Filters/Contents/FilterSelectorContents.component.tsx
@@ -3,6 +3,7 @@ import { withStyles, WithStyles } from 'capture-core-utils/styles';
 import { useFeature, FEATURES } from 'capture-core-utils/featuresSupport';
 import { MAX_OPTIONS_COUNT_FOR_OPTION_SET_CONTENTS, filterTypesObject } from '../filters.const';
 import { withButtons } from './withButtons';
+import { withMinCharsToSearchValidation } from './withMinCharsToSearchValidation';
 import {
     TextFilter,
     NumericFilter,
@@ -79,7 +80,7 @@ const useContents = ({
         }
 
         const TypeFilter = selectorContentsForTypes[type];
-        return [withButtons()(TypeFilter), false];
+        return [withButtons()(withMinCharsToSearchValidation()(TypeFilter)), false];
     }, [type, options]);
 
     if (EMPTY_ONLY_FILTER_TYPES.has(type) && !emptyValueFilterSupported) {