+++ src/core_modules/capture-core/components/ListView/Filters/Filters.component.tsx
@@ -150,7 +150,7 @@ const addAdditionalFiltersElements = (
     const mainButtonAdditionalFilters = additionalFilters && additionalFilters.find(filter => filter.mainButton);
     if (mainButtonAdditionalFilters) {
         const addToRemainingElements =
-        !filtersWithValueOnInit[mainButtonAdditionalFilters.id] && !userSelectedFilters[mainButtonAdditionalFilters.id];
+            !filtersWithValueOnInit[mainButtonAdditionalFilters.id] && !userSelectedFilters[mainButtonAdditionalFilters.id];
 
         if (addToRemainingElements) {
             const remainingElements =
@@ -220,31 +220,46 @@ const renderIndividualFilterButtons = ({
     onRemoveFilter: RemoveFilter;
     classes: any;
 }) => [...(filtersOnly || []), ...individualElementsArray]
-    .map(({ id, type, header, options, multiValueFilter, disabled, tooltipContent, mainButton, unique }: any) => (
-        <div
-            key={id}
-            data-test={`filter-button-container-${id}`}
-            className={classes.filterButtonContainer}
-        >
-            <FilterButton
-                data-test={`filter-button-${id}`}
-                itemId={id}
-                type={type}
-                title={header}
-                options={options}
-                disabled={disabled}
-                tooltipContent={tooltipContent}
-                multiValueFilter={multiValueFilter}
-                unique={unique}
-                onSetVisibleSelector={onSetVisibleSelector}
-                selectorVisible={id === visibleSelectorId}
-                onUpdateFilter={onUpdateFilter}
-                onClearFilter={onClearFilter}
-                isRemovable={mainButton}
-                onRemoveFilter={onRemoveFilter}
-            />
-        </div>
-    ),
+    .map(
+        ({
+            id,
+            type,
+            header,
+            options,
+            multiValueFilter,
+            disabled,
+            tooltipContent,
+            mainButton,
+            unique,
+            searchOperator,
+            minCharactersToSearch,
+        }: any) => (
+            <div
+                key={id}
+                data-test={`filter-button-container-${id}`}
+                className={classes.filterButtonContainer}
+            >
+                <FilterButton
+                    data-test={`filter-button-${id}`}
+                    itemId={id}
+                    type={type}
+                    title={header}
+                    options={options}
+                    disabled={disabled}
+                    tooltipContent={tooltipContent}
+                    multiValueFilter={multiValueFilter}
+                    unique={unique}
+                    searchOperator={searchOperator}
+                    minCharactersToSearch={minCharactersToSearch}
+                    onSetVisibleSelector={onSetVisibleSelector}
+                    selectorVisible={id === visibleSelectorId}
+                    onUpdateFilter={onUpdateFilter}
+                    onClearFilter={onClearFilter}
+                    isRemovable={mainButton}
+                    onRemoveFilter={onRemoveFilter}
+                />
+            </div>
+        ),
     );
 
 const renderRestButton = (
@@ -300,10 +315,10 @@ const FiltersPlain = memo<Props & WithStyles<typeof getStyles>>((props: Props & 
             splitBasedOnHasValueOnInit(validElementConfigs, filtersWithValueOnInit);
 
         const { fillUpElements, remainingElements: remainingElementsAfterFillUp } =
-        fillUpIndividualElements(
-            remainingElementsAfterInitSplit,
-            initValueElements.size + defaultFiltersOnlyCount,
-        );
+            fillUpIndividualElements(
+                remainingElementsAfterInitSplit,
+                initValueElements.size + defaultFiltersOnlyCount,
+            );
 
         const { remainingElements: remainingElementsWithShowInMoreFilters } = addShowInMoreFilters(
             remainingElementsAfterFillUp,
@@ -317,7 +332,7 @@ const FiltersPlain = memo<Props & WithStyles<typeof getStyles>>((props: Props & 
                 remainingElementsWithShowInMoreFilters, additionalFilters, filtersWithValueOnInit, userSelectedFilters);
 
         const { userSelectedElements, remainingElements } =
-        getUserSelectedElements(remainingElementsWithAdditionalFilters, userSelectedFilters);
+            getUserSelectedElements(remainingElementsWithAdditionalFilters, userSelectedFilters);
 
         const individualElementsArray =
             getIndividualElementsArray(