+++ src/core_modules/capture-core/components/WorkingLists/TrackerWorkingLists/Setup/hooks/useDefaultColumnConfig.ts
@@ -90,7 +90,17 @@ const getEventsMetaDataConfig = (programStage): Array<MetadataColumnConfig> => {
 };
 
 const getTEIMetaDataConfig = (attributes: Array<any>, orgUnitId: string | null | undefined): Array<MetadataColumnConfig> =>
-    attributes.map(({ id, displayInReports, type, name, formName, optionSet, searchable, unique }) => ({
+    attributes.map(({
+        id,
+        displayInReports,
+        type,
+        name,
+        formName,
+        optionSet,
+        searchable,
+        unique,
+        searchOperator,
+        minCharactersToSearch }) => ({
         id,
         visible: displayInReports,
         type,
@@ -99,6 +109,8 @@ const getTEIMetaDataConfig = (attributes: Array<any>, orgUnitId: string | null |
         multiValueFilter: !!optionSet || type === dataElementTypes.BOOLEAN,
         filterHidden: !(orgUnitId || searchable || unique),
         unique: Boolean(unique),
+        searchOperator,
+        minCharactersToSearch,
     }));
 
 const getDataValuesMetaDataConfig = (dataElements): Array<MetadataColumnConfig> =>